fix position of DLC (un)owned base game sign
currently the +/- sign used to indicate (un)owned base game is displayed way above the other icons and slightly to the left, which is caused by -4% top and left margin. changing them to 0% positions the sign nicely next to the DLC icon; 
of course with these changes, margin settings can just be removed completely.